// src/types/Slider.ts
export interface Slider {
    id: number;
    path: string;
    path_url: string;
    row_1_text: string;
    row_1_color: string;
    row_1_css: string;
    row_2_text: string;
    row_2_color: string;
    row_2_css: string;
    button_text: string;
    button_url: string;
    button_target: string;
    button_color: string;
    button_css: string;
    is_active: boolean;
}